requirements.txt
cloudevents==1.11.0
Flask==3.0.3
functions-framework==3.8.1
google-api-core==2.20.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
slack_sdk==3.33.2
